refactor test fixtures and add gcp function deployment to actions.
