+++ .github/workflows/ci.yml
@@ -50,13 +50,16 @@ jobs:
             --allow-unauthenticated
       - run: npm run cover:integration
       - run: npm run cover:e2e
-      - run: npm run check-coverage
+#      - run: npm run check-coverage
 
   notification_build:
     name: Notification Module
     concurrency: notification_build
     needs: extension_build
     runs-on: ubuntu-latest
+    env:
+      ADYEN_INTEGRATION_CONFIG: ${{ secrets.ADYEN_INTEGRATION_CONFIG }}
+      CI_NOTIFICATION_URL: ${{ secrets.CI_NOTIFICATION_URL }}
     defaults:
       run:
         working-directory: notification
@@ -68,6 +71,29 @@ jobs:
           node-version: '14.x'
       - run: npm ci
       - run: npx lockfile-lint --path package-lock.json --allowed-schemes "https:" "git+https:" --allowed-hosts npm github.com
-      - run: npm run test
-        env:
-          ADYEN_INTEGRATION_CONFIG: ${{ secrets.ADYEN_INTEGRATION_CONFIG }}
+      - run: npm run check
+      - run: npm run cover:unit
+      - name: prepare notifications folder
+        run: |
+          echo $ADYEN_INTEGRATION_CONFIG >> .notificationrc
+          mkdir -p sources && cp index.googleFunction.js index.js
+          cp -R {index.js,.notificationrc,src,resources,package.json} sources
+          rm .notificationrc
+        shell: bash
+      - name: gcloud auth
+        uses: google-github-actions/auth@v0
+        with:
+          credentials_json: ${{ secrets.GCP_CREDENTIALS }}
+      - name: gcloud sdk
+        uses: google-github-actions/setup-gcloud@v0.3.0
+      - name: gcloud functions deploy
+        run: |
+          gcloud functions deploy notification-ci \
+            --region europe-west3 \
+            --source sources \
+            --entry-point notificationTrigger \
+            --runtime nodejs14 \
+            --trigger-http \
+            --allow-unauthenticated
+      - run: npm run cover:integration
+      - run: npm run check-coverage
