+++ notification/test/test-utils.js
@@ -1,44 +1,17 @@
+const _ = require('lodash')
+const { address } = require('ip')
+const { hmacValidator } = require('@adyen/api-library')
 const config = require('../src/config/config')
 const concurrentModificationError = require('./resources/concurrent-modification-exception.json')
+const serverBuilder = require('../src/server')
+const { setupNotificationResources } = require('../src/setup')
+const payment = require('./resources/payment-draft.json')
 
 process.on('unhandledRejection', (reason) => {
   /* eslint-disable no-console */
   console.error('Unhandled Rejection:', reason)
   process.exit(1)
 })
-
-async function deleteResource(ctpClient, endpoint, item) {
-  const uri = ctpClient.builder[endpoint]
-  return ctpClient.delete(uri, item.id, item.version)
-}
-
-async function unpublish(ctpClient, product) {
-  const uri = ctpClient.builder.products
-  const actions = [
-    {
-      action: 'unpublish',
-    },
-  ]
-  const res = await ctpClient.update(uri, product.id, product.version, actions)
-  return res.body
-}
-
-async function deleteAllResources(ctpClient, endpoint, condition) {
-  let requestBuilder = ctpClient.builder[endpoint]
-
-  if (condition) requestBuilder = requestBuilder.where(condition)
-
-  return ctpClient.fetchBatches(requestBuilder, (items) =>
-    Promise.all(
-      items.map(async (item) => {
-        if (endpoint === 'products' && item.masterData.published)
-          item = await unpublish(ctpClient, item)
-
-        return deleteResource(ctpClient, endpoint, item)
-      })
-    )
-  )
-}
 
 let originalGetAdyenConfigFn
 
@@ -66,10 +39,109 @@ function buildMockErrorFromConcurrentModificaitonException() {
   return error
 }
 
+// node-fetch package doesn't support requests to localhost, therefore
+// we need to provide the IP behind localhost
+const localhostIp = address()
+
+async function startIT() {
+  await setupNotificationResources()
+  if (!process.env.CI) {
+    await setupLocalServer(8000)
+    return `http://${localhostIp}:8000`
+  } else {
+    return process.env.CI_NOTIFICATION_URL
+  }
+}
+
+let server
+async function setupLocalServer(testServerPort = 8000) {
+  server = serverBuilder.setupServer()
+  return new Promise((resolve) => {
+    server.listen(testServerPort, async () => {
+      resolve()
+    })
+  })
+}
+
+function stopIT() {
+  server.close()
+}
+
+const validator = new hmacValidator()
+
+function createNotificationPayload(
+  commercetoolsProjectKey,
+  adyenMerchantAccount,
+  paymentKey,
+  eventCode = 'AUTHORISATION',
+  pspReference = 'test_AUTHORISATION_1',
+  success = 'true'
+) {
+  const notification = {
+    live: 'false',
+    notificationItems: [
+      {
+        NotificationRequestItem: {
+          additionalData: {
+            'metadata.ctProjectKey': commercetoolsProjectKey,
+          },
+          amount: {
+            currency: 'EUR',
+            value: 10100,
+          },
+          eventCode,
+          eventDate: '2019-01-30T18:16:22+01:00',
+          merchantAccountCode: adyenMerchantAccount,
+          merchantReference: paymentKey,
+          operations: ['CANCEL', 'CAPTURE', 'REFUND'],
+          paymentMethod: 'visa',
+          pspReference,
+          success,
+        },
+      },
+    ],
+  }
+  const adyenConfig = config.getAdyenConfig(adyenMerchantAccount)
+  const notificationRequestItem =
+    notification.notificationItems[0].NotificationRequestItem
+  if (eventCode === 'REFUND') {
+    notificationRequestItem.additionalData['modification.action'] = 'refund'
+  } else if (eventCode === 'CANCEL_OR_REFUND') {
+    notificationRequestItem.additionalData['modification.action'] = 'cancel'
+  }
+  if (adyenConfig.enableHmacSignature) {
+    notificationRequestItem.additionalData.hmacSignature =
+      validator.calculateHmac(
+        notificationRequestItem,
+        adyenConfig.secretHmacKey
+      )
+  }
+
+  return notification
+}
+
+function ensurePayment(
+  ctpClient,
+  paymentKey,
+  commercetoolsProjectKey,
+  adyenMerchantAccount
+) {
+  const paymentDraft = _.cloneDeep(payment)
+  paymentDraft.key = paymentKey
+  paymentDraft.custom.fields = {
+    commercetoolsProjectKey,
+    adyenMerchantAccount,
+  }
+
+  return ctpClient.create(ctpClient.builder.payments, paymentDraft)
+}
+
 module.exports = {
-  unpublish,
-  deleteAllResources,
   overrideAdyenConfig,
   restoreAdyenConfig,
   buildMockErrorFromConcurrentModificaitonException,
+  startIT,
+  stopIT,
+  createNotificationPayload,
+  ensurePayment,
 }
